General improvements and bugfixes.
Added a saving-loader when saving values.
Fix loading of values and rendering of EditDisplay and its children to not lose state when navigating with unsaved values.
Fixed a bug where multiple open namespaces with the same selected key-name where showing both as selected.
Fixed unecessary loading and errorstates in displayReducer.
Switched formating-icon in editor to be more in line with the expected action.